backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-jose
passlib
python-multipart
langchain
langchain_community
langgraph
supabase
pytest
httpx
python-dotenv
bcrypt
redis
prometheus-client
prometheus-fastapi-instrumentator
email-validator

